==> 1_DataCleaning_metacritic.ipynb <==
import numpy as np 
import pandas as pd 
from langdetect import detect
import time
import pickle

# IMPORT

df_info = pd.read_csv('raw_data/metacritic_game_info.csv')
df_info.head(3)

df_comments = pd.read_csv('raw_data/metacritic_game_user_comments.csv')
df_comments.head(3)

df_info.shape, df_comments.shape

# merge

t1 = df_info.Title.unique()
t2 = df_comments.Title.unique()
len(t1),len(t2)

df_con = pd.merge(df_info, df_comments, on='Title', how='right')

df_comments.join(df_info, on='Title')

df_con.shape

df_con.isna().any()

print(f'shape before: {df_con.shape}')
df_con.dropna(inplace=True)
df_con.drop_duplicates(inplace=True)
print(f'shape after: {df_con.shape}')

df_con

df_comments.Userscore.value_counts(normalize=True)

# Metacritic Game Info

# remove the extra index

df_info = df_info.drop(columns='Unnamed: 0')
df_info.head(1)

# remove duplicates

print(f'{df_info.duplicated().sum()} duplicates have been removed.')

df_info.drop_duplicates(inplace=True)

df_info.nunique()

- year

# remove 'not specified' in column Year
# 4 of them

t = df_info[df_info['Year'] == 'not specified'].shape[0]
print(f'{t} nans have been removed.')

drop_index = df_info[df_info['Year'] == 'not specified'].index
df_info.drop(drop_index, inplace=True)

df_info.Year.unique()

df_info.Year.value_counts().sort_index(ascending=False)

- Publisher

df_info['Publisher'].value_counts(normalize=True)

ohe_list_column(df_info,Year)

df_info.columns

- genre

def find_uniques(df,feature='feature',sep=';'):
    df[feature] = df[feature].apply(lambda x: x.split(sep))

    genre_list = []

    df[feature].apply(lambda x: [genre_list.append(g) for g in x if g not in genre_list])

    return genre_list

genres_list = find_uniques(df_info, 'Genre', ';')
genres_list, len(genres_list)

def ohe_list_column(df,column):
    return df.drop(column, 1).join(df_info[column].str.join('|').str.get_dummies())

df_info = ohe_list_column(df_info,'Genre')

df_info

- No_Players

df_info['No_Players'].value_counts()

df_info[df_info['No_Players'].isna() == True] = 'not specified'

- metascore

df_info['Avg_Userscore'].unique()

df_info.columns

- quick export

# a quick export for easy remsy

df_info.to_csv(r'processed_remsy\metacritic_info_remsy.csv', index = False)

# Metacritic Review

# remove the extra index

df_comments = df_comments.drop(columns='Unnamed: 0')
df_comments.head(1)

df_comments.nunique()

df_comments.shape

df_comments.drop_duplicates(inplace=True)
df_comments.dropna(inplace=True)

df_comments.shape

# langauage detect

df_comments_5000 = df_comments.sample(5000)

def detect_lang(t):
    try:
        lang = detect(t)
    except:
        lang = 'error'

    return lang


start = time.time()

df_comments_5000['lang'] = df_comments_5000['Comment'].apply(detect_lang)

end = time.time()
print(f'this function took {end-start} seconds.')

df_comments_5000

df_comments_5000[df_comments_5000.lang=='error']

df_comments_5000['lang'].value_counts(normalize=True)

start = time.time()

df_comments['lang'] = df_comments['Comment'].apply(detect_lang)

end = time.time()
print(f'this function took {end-start} seconds.')

pickle

# df_comments.to_pickle('processed_df/metacritic_comment_langdetected.pkl')

df_comments = pickle.load(open('processed_df/metacritic_comment_langdetected.pkl','rb'))

df_comments

df_comments['lang'].value_counts(normalize=True)

# EXPORT

df_comments.columns

df_comments.shape

df_comments.Userscore.unique()

# remove non-english reviews
print(f'shape before: {df_comments.shape}')
df_comments = df_comments[df_comments['lang'] == 'en']
print(f'shape after removed non-english: {df_comments.shape}')

- remsy

col_sel = ['Username', 'Userscore', 'Title', 'Comment']
df_comments_remsy = df_comments[col_sel]

df_comments_remsy.shape

# replace userscore 0.0 with nan
# assume 0.0 means not rated
# this df is for NLP EDA and modeling, doesn't need missing values

df_comments_remsy.loc[df_comments_remsy['Userscore'] == 0, 'Userscore'] = np.NAN

df_comments_remsy.Userscore.unique()

df_comments_remsy.to_csv(r'processed_remsy\metacritic_comments_remsy.csv', index = False)

- NLP

# replace userscore 0.0 with nan
# assume 0.0 means not rated
# this df is for NLP EDA and modeling, doesn't need missing values

df_comments.loc[df_comments['Userscore'] == 0, 'Userscore'] = np.NAN


df_comments.Userscore.unique()

df_comments_nlp = df_comments.copy()

print(f'shape before: {df_comments_nlp.shape}')
df_comments_nlp.drop_duplicates(inplace=True)
print(f'shape after: {df_comments_nlp.shape}')

df_need_proc = df_comments_nlp[df_comments_nlp['Userscore'].isna() == True]
df_need_proc.to_csv(r'processed_df\df_need_proc.csv', index = False)

df_comments_nlp[df_comments_nlp['Userscore'].isna() == False].to_csv(r'processed_df\metacritic_comments_nlp.csv', index = False)
