==> 1_DataCleaning_SteamStore.ipynb <==
import numpy as np 
import pandas as pd
import time
import pickle

# IMPORT

df1 = pd.read_csv('raw_data\Steam Store Games\steam.csv')
df1.head(3)

df2 = pd.read_csv('raw_data\Steam Store Games\steam_description_data.csv')
df2.head(3)

df3 = pd.read_csv('raw_data\Steam Store Games\steamspy_tag_data.csv')
df3.head(3)

df4 = pd.read_csv('raw_data/Steam_Game_Review_Dataset/train_gr/train.csv')
df4.head(3)

df5 = pd.read_csv('raw_data/Steam_Game_Review_Dataset/train_gr/game_overview.csv')
df5.head(3)

- df1

df1.sample(3)

df1.nunique()

df1.columns

seperating list in columns

def find_uniques(df,feature='feature',sep=';'):
    df[feature] = df[feature].apply(lambda x: x.split(sep))

    genre_list = []

    df[feature].apply(lambda x: [genre_list.append(g) for g in x if g not in genre_list])

    return genre_list

genres_list = find_uniques(df1, 'genres', ';')
genres_list, len(genres_list)

categories_list = find_uniques(df1,'categories',';')
categories_list, len(categories_list)

df1.columns

platform_list = find_uniques(df1,'platforms',';')
platform_list, len(platform_list)

# create genres and categories one hot encode
df1 = df1.drop('genres', 1).join(df1.genres.str.join('|').str.get_dummies())

df1 = df1.drop('categories', 1).join(df1.categories.str.join('|').str.get_dummies())

df1 = df1.drop('platforms', 1).join(df1.platforms.str.join('|').str.get_dummies())

# drop some unnecessary columns
df1.drop(columns=['english','steamspy_tags','owners'], inplace=True)

df1

df1.columns

print(f'shape before: {df1.shape}')
df1.dropna(inplace=True)
df1.drop_duplicates(inplace=True)
print(f'shape after drop NaN and duplicates: {df1.shape}')

# EXPORT

- NLP

col_sel = ['title', 'user_review', 'user_suggestion']
df_steam_nlp = df4[col_sel]

df_steam_nlp.shape

df_steam_nlp.dropna(inplace=True)
df_steam_nlp.drop_duplicates(inplace=True)

df_steam_nlp.shape

df_steam_nlp.to_csv(r'D:\Flatiron\DS_P4\processed_df\steam_review_nlp.csv', index = False)

- recommendation system

df1.to_pickle('processed_df/steam_remsy.pkl')
# df of steam for recommendation system, ONE for genres and categories
