==> 2_EDA_Remsy_Metacritic.ipynb <==
import pandas as pd
import numpy as np
from copy import deepcopy
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import Ridge
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error

df = pd.read_csv('processed_remsy\metacritic_info_remsy.csv')
df

df_comments = pd.read_csv('raw_data/metacritic_game_user_comments.csv')
df_comments.head(3)

# remove the extra index

df_comments = df_comments.drop(columns='Unnamed: 0')

df_comments.nunique()

df = df.set_index('Title')

X = df.drop(columns=['Year','Publisher','Platform','No_Players'])
y = df['Avg_Userscore']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.15, random_state = 42)

X_train

# define the model
steps = [('scaler', StandardScaler()), ('rr', Ridge(alpha = 0.05)) ]
pipe_mod = Pipeline(steps)

# fir the model to the training data
pipe_mod.fit(X_train, y_train)
