==> 2_EDA_remsy_steam.ipynb <==
# standard packages for data analysis
import numpy as np
import pandas as pd
import pickle
import time

#visualization packages
import seaborn as sns
import matplotlib.pyplot as plt

df = pickle.load(open('processed_df/steam_remsy.pkl','rb'))
df.head(3)

df2 = pickle.load(open('processed_df/preprocessed_nlp_5p.pkl','rb'))
df.head(3)
