==> 3_Modelling_Metacritic_NLP_XGB.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import nltk
from nltk.tokenize import regexp_tokenize, word_tokenize, RegexpTokenizer
from nltk.corpus import stopwords, wordnet
from nltk import pos_tag
from nltk.stem import WordNetLemmatizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

from xgboost import XGBRegressor
from xgboost import XGBClassifier

from sklearn.model_selection import StratifiedShuffleSplit

from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score

import pickle
import time

# IMPORT

# df = pickle.load(open('processed_df\preprocessed_nlp_5p.pkl','rb'))
# df.head()

df = pd.read_csv('processed_df\preprocessed_nlp_30p.csv')
df.head()

df100 = pd.read_csv('processed_df\preprocessed_nlp_100p.csv')
df100.head()

print(f'shape before: {df.shape}')
df_new = df[df['Comment'] != '']
df_new.drop_duplicates(inplace=True)
df_new.dropna(inplace=True)
print(f'shape after: {df_new.shape}')


print(f'shape before: {df100.shape}')
df100_new = df100[df100['Comment'] != '']
df100_new.drop_duplicates(inplace=True)
df100_new.dropna(inplace=True)
print(f'shape after: {df100_new.shape}')


# Train test split

X = df_new['Comment']
y = df_new['Userscore']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

X100 = df100_new['Comment']
y100 = df100_new['Userscore']

X100_train, X100_test, y100_train, y100_test = train_test_split(X100, y100, test_size = 0.3)

# XGBoost

- baseline, 30% data, 160 vol

steps = [('tfidf', TfidfVectorizer(max_df=0.95, min_df=0.05)), ('XGBoost', XGBRegressor())]
xgb_pipeline = Pipeline(steps)

xgb_pipeline.fit(X_train, y_train)

y_pred = xgb_pipeline.predict(X_test)
print(f'test MSE {mean_squared_error(y_test, y_pred)}')
print(f'test MAE {mean_absolute_error(y_test, y_pred)}')
print(f'test r2 {r2_score(y_test, y_pred)}')

print('  ')

y_pred = xgb_pipeline.predict(X_train)
print(f'train MSE {mean_squared_error(y_train, y_pred)}')
print(f'train MAE {mean_absolute_error(y_train, y_pred)}')
print(f'train r2 {r2_score(y_train, y_pred)}')

xgb_pipeline['XGBoost'].get_params()

- baseline, 30% data, 1000 vol

steps = [('tfidf', TfidfVectorizer(max_features=1000)), ('XGBoost', XGBRegressor())]
xgb_pipeline_m = Pipeline(steps)

start = time.time()

xgb_pipeline_m.fit(X_train, y_train)

end = time.time()
print(f'{end-start} secs slipped..')

y_pred = xgb_pipeline_m.predict(X_test)
print(f'test MSE {mean_squared_error(y_test, y_pred)}')
print(f'test MAE {mean_absolute_error(y_test, y_pred)}')
print(f'test r2 {r2_score(y_test, y_pred)}')

print('  ')

y_pred = xgb_pipeline_m.predict(X_train)
print(f'train MSE {mean_squared_error(y_train, y_pred)}')
print(f'train MAE {mean_absolute_error(y_train, y_pred)}')
print(f'train r2 {r2_score(y_train, y_pred)}')

- baseline, 30% data, 5000 vol

steps = [('tfidf', TfidfVectorizer(max_features=5000)), ('XGBoost', XGBRegressor())]
xgb_pipeline_2 = Pipeline(steps)

start = time.time()
xgb_pipeline_2.fit(X_train, y_train)
end = time.time()
print(f'training used {end-start} secs')
print('  ')

y_pred = xgb_pipeline_2.predict(X_test)
print(f'test MSE {mean_squared_error(y_test, y_pred)}')
print(f'test MAE {mean_absolute_error(y_test, y_pred)}')
print(f'test r2 {r2_score(y_test, y_pred)}')
print('  ')

y_pred = xgb_pipeline_2.predict(X_train)
print(f'train MSE {mean_squared_error(y_train, y_pred)}')
print(f'train MAE {mean_absolute_error(y_train, y_pred)}')
print(f'train r2 {r2_score(y_train, y_pred)}')


- baseline, 100% data, 5000 vol

steps = [('tfidf', TfidfVectorizer(max_features=5000)), ('XGBoost', XGBRegressor())]
xgb_pipeline_3 = Pipeline(steps)

start = time.time()
xgb_pipeline_3.fit(X100_train, y100_train)
end = time.time()
print(f'training used {end-start} secs')
print('  ')

y_pred = xgb_pipeline_3.predict(X100_test)
print(f'test MSE {mean_squared_error(y100_test, y_pred)}')
print(f'test MAE {mean_absolute_error(y100_test, y_pred)}')
print(f'test r2 {r2_score(y100_test, y_pred)}')
print('  ')

y_pred = xgb_pipeline_3.predict(X100_train)
print(f'train MSE {mean_squared_error(y100_train, y_pred)}')
print(f'train MAE {mean_absolute_error(y100_train, y_pred)}')
print(f'train r2 {r2_score(y100_train, y_pred)}')


best_model = xgb_pipeline_3
best_model

pickle.dump(best_model, open('models/xgb_baseline.pkl', 'wb'))

loaded_model = pickle.load(open('models/xgb_baseline.pkl', 'rb'))

y_pred = loaded_model.predict(X100_test)
print(f'test r2 {r2_score(y100_test, y_pred)}')

# GRID SEARCH
