==> 4_RecommendationSystem_XGB_Metacritic.ipynb <==
import pandas as pd
import numpy as np
from copy import deepcopy
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.linear_model import Ridge
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error

from surprise import Dataset, Reader
from surprise import SVD
from surprise import accuracy
from surprise.model_selection import cross_validate, train_test_split
from surprise.prediction_algorithms import knns
from surprise.similarities import cosine, msd, pearson
from surprise.model_selection import GridSearchCV
from surprise.prediction_algorithms import KNNWithMeans, KNNBasic, KNNBaseline

import time
import pickle

# NLP modules we will use for text normalization
import re #regex 
import nltk # the natural language toolkit
from nltk.tokenize import word_tokenize
from nltk import WordNetLemmatizer
from nltk.corpus import stopwords
from nltk.corpus import wordnet
from nltk import pos_tag


# import

df = pd.read_csv('processed_remsy\metacritic_comments_remsy.csv')
df

XGB_model = pickle.load(open('models/xgb_baseline.pkl', 'rb'))

# SENTIMENTAL IMPUTE

# def a function to process_text for TFIDF
def process_text(text, min_length):

    # drop non-english words
    words = set(nltk.corpus.words.words())
    # " ".join(w for w in nltk.wordpunct_tokenize(text) if w.lower() in words or not w.isalpha())

    # get common stop words that we'll remove during tokenization/text normalization
    stop_words = stopwords.words('english')

    #initialize lemmatizer
    wnl = WordNetLemmatizer()

    # helper function to change nltk's part of speech tagging to a wordnet format.
    def pos_tagger(nltk_tag):
        if nltk_tag.startswith('J'):
            return wordnet.ADJ
        elif nltk_tag.startswith('V'):
            return wordnet.VERB
        elif nltk_tag.startswith('N'):
            return wordnet.NOUN
        elif nltk_tag.startswith('R'):
            return wordnet.ADV
        else:         
            return None
   

    # lower case everything
    text_lower = text.lower()
    
    # remove stop words and punctuations 
    text_norm = [x for x in word_tokenize(text_lower) if ((x.isalpha()) & (x not in stop_words))]

    #  POS detection on the result will be important in telling Wordnet's lemmatizer how to lemmatize
    
    # creates list of tuples with tokens and POS tags in wordnet format
    wordnet_tagged = list(map(lambda x: (x[0], pos_tagger(x[1])), pos_tag(text_norm))) 

    # cutoff for tokenized length
    if len(wordnet_tagged) <= min_length:
        return ''
    else:
         # rejoins lemmatized sentence 
         text_norm = " ".join([wnl.lemmatize(x[0], x[1]) for x in wordnet_tagged if x[1] is not None])
         return text_norm

# subset the dataset p * origin
# then do text process
# lastly, get rid of null, na and duplicates, because of the limit of length

def subset_and_process(p, df, colname):
    
    start = time.time()

    sample_size = int(df.shape[0]*p)
    print(f'subset size: {sample_size}')
    print(f'full dataset size: {df.shape[0]}')
    print(f'{p*100}%')

    df_sub = df.sample(sample_size,random_state=2333) 
    df_sub[colname].value_counts(normalize=True)

    start = time.time()
    df_sub[colname] = df_sub[colname].apply(process_text, min_length = 10)
    end = time.time()
    print(f'{end-start} seconds slipped.')

    df_new = df_sub[df_sub[colname] != '']
    df_new.drop_duplicates(inplace=True)

    print(f'output dataset size: {df_new.shape[0]}')
    
    end = time.time()
    print(f'{end-start} secs slipped..')

    return df_new

df_nan = df[df['Userscore'].isna()==True]
df_nan

# select the review with no score
df_nan_proc = subset_and_process(1, df_nan, 'Comment')

# df_nan_proc.to_csv(r'processed_df\df_nan_proc.csv', index = False)


df_nan_proc = pd.read_csv('processed_df\df_nan_proc.csv')
df_nan_proc

# Getting the data ready for Surprise

data = df[['Username', 'Title', 'Userscore']]

reader = Reader(line_format='user item rating', sep=',')
data = Dataset.load_from_df(data, reader=reader)

# train test split

trainset, testset = train_test_split(data, test_size=.2, random_state = 2333)

print('Type trainset :',type(trainset),'\n')
print('Type testset :',type(testset))

print(len(testset))
print(testset[0])

print('Number of users: ', trainset.n_users, '\n')
print('Number of items: ', trainset.n_items, '\n')

KNN

- baseline

sim_cos = {'name':'cosine', 'user_based':False}

basic = knns.KNNBasic(sim_options=sim_cos)
basic.fit(trainset)

basic.sim

predictions = basic.test(testset)

print(accuracy.rmse(predictions))

# SVD

- baseline

svd = SVD()
svd.fit(trainset)

predictions = svd.test(testset)
accuracy.rmse(predictions)

- GridSearch

start = time.time()

params = {'n_factors': [20, 50, 100], 'reg_all': [0.02, 0.05, 0.1]}
g_s_svd = GridSearchCV(SVD, param_grid=params, n_jobs=-1, cv=5)
g_s_svd.fit(data)

end = time.time()
print(f'time spent {end-start}')

print(g_s_svd.best_score)
print(g_s_svd.best_params)

# slightly better

- predict

# use sentiment analyzer to impute nan score

- impute

df_nan_proc

df_nan_proc['Userscore'] = XGB_model.predict(df_nan_proc['Comment'])

df_nan_proc

frames = [df[df['Userscore'].isna()==False], df_nan_proc]
df_con = pd.concat(frames)

- prepare data

data_con = df_con[['Username', 'Title', 'Userscore']]

reader = Reader(line_format='user item rating', sep=',')
data_con = Dataset.load_from_df(data_con, reader=reader)

# train test split

trainset, testset = train_test_split(data_con, test_size=.2, random_state = 2333)

- KNN

sim_cos = {'name':'cosine', 'user_based':False}

basic = knns.KNNBasic(sim_options=sim_cos)
basic.fit(trainset)

predictions = basic.test(testset)
print(accuracy.rmse(predictions))

- SVD

svd = SVD()
svd.fit(trainset)

predictions = svd.test(testset)
accuracy.rmse(predictions)
